update
ShopCart [add, modify, delete]

+++ CoffeeTime/ActivityTableViewCell.h
@@ -2,7 +2,7 @@
 //  ActivityTableViewCell.h
 //  CoffeeTime
 //
-//  Created by fule on 15/7/7.
+//  Created by Adam on 15/7/7.
 //  Copyright (c) 2015年 fule. All rights reserved.
 //
 
@@ -11,7 +11,7 @@
 
 @protocol ActivityTableViewCellDelete <NSObject>
 
-- (void)btnClick:(UITableViewCell *)cell andFlag:(int)flag;
+- (void)btnClick:(UITableViewCell *)cell andFlag:(NSInteger)flag;
 
 @end
 
@@ -23,7 +23,8 @@
 @property (strong, nonatomic) IBOutlet UIButton *subBtn;
 @property (strong, nonatomic) IBOutlet UILabel *countLabel;
 
-@property(assign,nonatomic)id<ActivityTableViewCellDelete>delegate;
+@property (assign,nonatomic) id<ActivityTableViewCellDelete> delegate;
+
 -(void)addTheValue:(ActivityModel *)activiyModel indexPath:(NSIndexPath*)indexPath;
 
 @end
